feat: [STK-62]  Rename files
